Remove reduntant state setting functions
