+++ modules/gltools/src/index.js
@@ -4,17 +4,10 @@ export {default as polyfillContext} from './polyfill/polyfill-context';
 
 // unified parameter APIs
 export {
-  getParameter,
   getParameters,
-  setParameter,
-  resetParameters,
-  getModifiedParameters
+  setParameters,
+  resetParameters
 } from './state-tracker/unified-parameter-api/unified-parameter-api';
-
-export {
-  // Support function style parameter keys
-  setParameters // TODO - setParameter should also support function style keys?
-} from './state-tracker/unified-parameter-api/set-parameters';
 
 // state tracking
 export {
