+++ modules/gltools/test/state-tracker/state-tracking/track-context-state.spec.js
@@ -5,10 +5,9 @@ import {
   trackContextState,
   pushContextState,
   popContextState,
-  getParameter,
+  getParameters,
   setParameters,
-  resetParameters,
-  getModifiedParameters
+  resetParameters
 } from '@luma.gl/gltools';
 
 import {
@@ -52,7 +51,7 @@ test('WebGLState#push & pop', t => {
 
   // Verify default values.
   for (const key in GL_PARAMETER_DEFAULTS) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       GL_PARAMETER_DEFAULTS[key],
@@ -65,7 +64,7 @@ test('WebGLState#push & pop', t => {
   // Set custom values and verify.
   setParameters(gl, ENUM_STYLE_SETTINGS_SET1);
   for (const key in ENUM_STYLE_SETTINGS_SET1) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       ENUM_STYLE_SETTINGS_SET1[key],
@@ -78,7 +77,7 @@ test('WebGLState#push & pop', t => {
   // Set custom values and verify
   setParameters(gl, ENUM_STYLE_SETTINGS_SET2);
   for (const key in ENUM_STYLE_SETTINGS_SET2) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       ENUM_STYLE_SETTINGS_SET2[key],
@@ -90,7 +89,7 @@ test('WebGLState#push & pop', t => {
   popContextState(gl);
 
   for (const key in ENUM_STYLE_SETTINGS_SET1) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       ENUM_STYLE_SETTINGS_SET1[key],
@@ -101,7 +100,7 @@ test('WebGLState#push & pop', t => {
   popContextState(gl);
 
   for (const key in GL_PARAMETER_DEFAULTS) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       GL_PARAMETER_DEFAULTS[key],
@@ -119,7 +118,7 @@ test('WebGLState#gl API', t => {
 
   // Verify default values.
   for (const key in GL_PARAMETER_DEFAULTS) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       GL_PARAMETER_DEFAULTS[key],
@@ -142,7 +141,7 @@ test('WebGLState#gl API', t => {
   for (const key in ENUM_STYLE_SETTINGS_SET1) {
     // Skipping composite setters
     if (typeof GL_PARAMETER_SETTERS[key] !== 'string') {
-      const value = getParameter(gl, key);
+      const value = getParameters(gl, [key])[key];
       t.deepEqual(
         value,
         ENUM_STYLE_SETTINGS_SET1[key],
@@ -154,7 +153,7 @@ test('WebGLState#gl API', t => {
   popContextState(gl);
 
   for (const key in GL_PARAMETER_DEFAULTS) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       GL_PARAMETER_DEFAULTS[key],
@@ -173,65 +172,51 @@ test('WebGLState#intercept gl calls', t => {
   pushContextState(gl);
 
   gl.blendEquation(gl.FUNC_SUBTRACT, gl.FUNC_SUBTRACT);
-  t.is(getParameter(gl, gl.BLEND_EQUATION_RGB), gl.FUNC_SUBTRACT, 'direct gl call is tracked');
+  t.is(
+    getParameters(gl, [gl.BLEND_EQUATION_RGB])[gl.BLEND_EQUATION_RGB],
+    gl.FUNC_SUBTRACT,
+    'direct gl call is tracked'
+  );
 
   gl.blendFunc(gl.ONE, gl.ONE);
-  t.is(getParameter(gl, gl.BLEND_SRC_RGB), gl.ONE, 'direct gl call is tracked');
+  t.is(
+    getParameters(gl, [gl.BLEND_SRC_RGB])[gl.BLEND_SRC_RGB],
+    gl.ONE,
+    'direct gl call is tracked'
+  );
 
   gl.stencilMask(8);
-  t.is(getParameter(gl, gl.STENCIL_WRITEMASK), 8, 'direct gl call is tracked');
+  t.is(
+    getParameters(gl, [gl.STENCIL_WRITEMASK])[gl.STENCIL_WRITEMASK],
+    8,
+    'direct gl call is tracked'
+  );
 
   gl.stencilFunc(gl.NEVER, 0, 1);
-  t.is(getParameter(gl, gl.STENCIL_FUNC), gl.NEVER, 'direct gl call is tracked');
+  t.is(
+    getParameters(gl, [gl.STENCIL_FUNC])[gl.STENCIL_FUNC],
+    gl.NEVER,
+    'direct gl call is tracked'
+  );
 
   gl.stencilOp(gl.KEEP, gl.ZERO, gl.REPLACE);
-  t.is(getParameter(gl, gl.STENCIL_PASS_DEPTH_FAIL), gl.ZERO, 'direct gl call is tracked');
+  t.is(
+    getParameters(gl, [gl.STENCIL_PASS_DEPTH_FAIL])[gl.STENCIL_PASS_DEPTH_FAIL],
+    gl.ZERO,
+    'direct gl call is tracked'
+  );
 
   popContextState(gl);
 
   // Verify default values.
   for (const key in GL_PARAMETER_DEFAULTS) {
-    const value = getParameter(gl, key);
+    const value = getParameters(gl, [key])[key];
     t.deepEqual(
       value,
       GL_PARAMETER_DEFAULTS[key],
       `got expected value ${stringifyTypedArray(value)}`
     );
   }
-
-  t.end();
-});
-
-test('WebGLState#getModifiedParameters', t => {
-  const {gl} = fixture;
-
-  resetParameters(gl);
-
-  let modifiedParameters = getModifiedParameters(gl);
-
-  t.deepEqual(modifiedParameters, {}, 'Nothing is changed');
-
-  pushContextState(gl);
-
-  // Set custom values and verify
-  setParameters(gl, ENUM_STYLE_SETTINGS_SET1);
-  modifiedParameters = getModifiedParameters(gl);
-
-  for (const key in ENUM_STYLE_SETTINGS_SET1) {
-    t.deepEqual(
-      modifiedParameters[key],
-      ENUM_STYLE_SETTINGS_SET1[key],
-      `got expected value ${stringifyTypedArray(modifiedParameters[key])} for key: ${key}`
-    );
-  }
-
-  for (const key in modifiedParameters) {
-    if (!(key in ENUM_STYLE_SETTINGS_SET1)) {
-      t.fail('should not contain keys that are not changed');
-    }
-  }
-
-  popContextState(gl);
 
   t.end();
 });
